backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
django-filter==23.3
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
PyMySQL==1.1.1
requests==2.32.3
